+++ Engine/src/Core/ImGuiLayer.h
@@ -4,7 +4,7 @@ namespace Engine {
 
 	class ImGuiLayer
 	{
-		static void Init(int windowHandle);
+		static void Init(int64_t windowHandle);
 		static void ShutDown();
 		friend class ModuleCore;
